requirements.txt
feedparser
requests
beautifulsoup4
lxml
trafilatura
PyYAML
jinja2
python-dateutil
pandas
numpy
ruptures
matplotlib
tqdm
